state: re-order funcs. fix tests
