+++ state/execution_test.go
@@ -23,64 +23,6 @@ var (
 	nTxsPerBlock = 10
 )
 
-func TestValidateBlock(t *testing.T) {
-	state := state()
-	state.SetLogger(log.TestingLogger())
-
-	// proper block must pass
-	block := makeBlock(state, 1)
-	err := state.ValidateBlock(block)
-	require.NoError(t, err)
-
-	// wrong chain fails
-	block = makeBlock(state, 1)
-	block.ChainID = "not-the-real-one"
-	err = state.ValidateBlock(block)
-	require.Error(t, err)
-
-	// wrong height fails
-	block = makeBlock(state, 1)
-	block.Height += 10
-	err = state.ValidateBlock(block)
-	require.Error(t, err)
-
-	// wrong total tx fails
-	block = makeBlock(state, 1)
-	block.TotalTxs += 10
-	err = state.ValidateBlock(block)
-	require.Error(t, err)
-
-	// wrong blockid fails
-	block = makeBlock(state, 1)
-	block.LastBlockID.PartsHeader.Total += 10
-	err = state.ValidateBlock(block)
-	require.Error(t, err)
-
-	// wrong app hash fails
-	block = makeBlock(state, 1)
-	block.AppHash = []byte("wrong app hash")
-	err = state.ValidateBlock(block)
-	require.Error(t, err)
-
-	// wrong consensus hash fails
-	block = makeBlock(state, 1)
-	block.ConsensusHash = []byte("wrong consensus hash")
-	err = state.ValidateBlock(block)
-	require.Error(t, err)
-
-	// wrong results hash fails
-	block = makeBlock(state, 1)
-	block.LastResultsHash = []byte("wrong results hash")
-	err = state.ValidateBlock(block)
-	require.Error(t, err)
-
-	// wrong validators hash fails
-	block = makeBlock(state, 1)
-	block.ValidatorsHash = []byte("wrong validators hash")
-	err = state.ValidateBlock(block)
-	require.Error(t, err)
-}
-
 func TestApplyBlock(t *testing.T) {
 	cc := proxy.NewLocalClientCreator(dummy.NewDummyApplication())
 	proxyApp := proxy.NewAppConns(cc, nil)
@@ -88,15 +30,16 @@ func TestApplyBlock(t *testing.T) {
 	require.Nil(t, err)
 	defer proxyApp.Stop()
 
-	state := state()
-	state.SetLogger(log.TestingLogger())
+	state, stateDB := state(), dbm.NewMemDB()
 
-	block := makeBlock(state, 1)
-
-	err = state.ApplyBlock(types.NopEventBus{}, proxyApp.Consensus(),
-		block, block.MakePartSet(testPartSize).Header(),
+	blockExec := NewBlockExecutor(stateDB, log.TestingLogger(),
+		types.NopEventBus{}, proxyApp.Consensus(),
 		types.MockMempool{}, types.MockEvidencePool{})
 
+	block := makeBlock(state, 1)
+	blockID := types.BlockID{block.Hash(), block.MakePartSet(testPartSize).Header()}
+
+	state, err = blockExec.ApplyBlock(state, blockID, block)
 	require.Nil(t, err)
 
 	// TODO check state and mempool
@@ -112,15 +55,14 @@ func TestBeginBlockAbsentValidators(t *testing.T) {
 	defer proxyApp.Stop()
 
 	state := state()
-	state.SetLogger(log.TestingLogger())
 
 	// there were 2 validators
-	val1PrivKey := crypto.GenPrivKeyEd25519()
+	/*val1PrivKey := crypto.GenPrivKeyEd25519()
 	val2PrivKey := crypto.GenPrivKeyEd25519()
 	lastValidators := types.NewValidatorSet([]*types.Validator{
 		types.NewValidator(val1PrivKey.PubKey(), 10),
 		types.NewValidator(val2PrivKey.PubKey(), 5),
-	})
+	})*/
 
 	prevHash := state.LastBlockID.Hash
 	prevParts := types.PartSetHeader{}
@@ -141,7 +83,7 @@ func TestBeginBlockAbsentValidators(t *testing.T) {
 		lastCommit := &types.Commit{BlockID: prevBlockID, Precommits: tc.lastCommitPrecommits}
 
 		block, _ := state.MakeBlock(2, makeTxs(2), lastCommit)
-		_, err = ExecCommitBlock(proxyApp.Consensus(), block, log.TestingLogger(), lastValidators)
+		_, err = ExecCommitBlock(proxyApp.Consensus(), block, log.TestingLogger())
 		require.Nil(t, err, tc.desc)
 
 		// -> app must receive an index of the absent validator
@@ -159,8 +101,8 @@ func makeTxs(height int64) (txs []types.Tx) {
 	return txs
 }
 
-func state() *State {
-	s, _ := MakeGenesisState(dbm.NewMemDB(), &types.GenesisDoc{
+func state() State {
+	s, _ := MakeGenesisState(&types.GenesisDoc{
 		ChainID: chainID,
 		Validators: []types.GenesisValidator{
 			{privKey.PubKey(), 10000, "test"},
@@ -170,7 +112,7 @@ func state() *State {
 	return s
 }
 
-func makeBlock(state *State, height int64) *types.Block {
+func makeBlock(state State, height int64) *types.Block {
 	block, _ := state.MakeBlock(height, makeTxs(state.LastBlockHeight), new(types.Commit))
 	return block
 }
